+++ tests/bmp/test_frr_bmp_sanity.py
@@ -1,5 +1,6 @@
 import pytest
-from tests.common.helpers.bgp import run_bgp_facts
+from tests.common.helpers.monit import check_monit_expected_container_logging
+from tests.common.utilities import wait_until
 from bmp.helper import enable_bmp_feature, disable_bmp_feature
 
 pytestmark = [
@@ -8,8 +9,10 @@ pytestmark = [
 ]
 
 
-def test_frr_bmp_bgp_facts(duthosts, enum_frontend_dut_hostname, enum_asic_index):
+def test_frr_bmp_monit_log(duthosts, enum_frontend_dut_hostname, enum_asic_index):
     duthost = duthosts[enum_frontend_dut_hostname]
     disable_bmp_feature(duthost)
-    run_bgp_facts(duthost, enum_asic_index)
+
+    wait_until(180, 60, 0, check_monit_expected_container_logging, duthost)
+
     enable_bmp_feature(duthost)
