+++ tests/monit/test_monit_status.py
@@ -9,6 +9,7 @@ import random
 from tests.common.utilities import wait_until
 from tests.common.helpers.assertions import pytest_assert
 from tests.common.helpers.assertions import pytest_require
+from tests.common.helpers.monit import check_monit_expected_container_logging
 
 logger = logging.getLogger(__name__)
 
@@ -75,19 +76,6 @@ def check_monit_last_output(duthost):
             return "/usr/bin/lldpmgrd' is not running in host" in monit_last_output
     else:
         return False
-
-
-def check_monit_expected_container_logging(duthost):
-    """Checks whether alerting message appears as syslog if
-    there is unexpected container not running.
-    Args:
-        duthost: An AnsibleHost object of DuT.
-    Returns:
-        None.
-    """
-    syslog_output = duthost.command("sudo grep 'ERR monit' /var/log/syslog")["stdout"]
-    pytest_assert("Expected containers not running" not in syslog_output,
-                  f"Expected containers not running found in syslog. Output was:\n{syslog_output}")
 
 
 def test_monit_status(duthosts, enum_rand_one_per_hwsku_frontend_hostname):
